# docs/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Dozlab API
  description: REST API for Dozlab learning platform microservice
  version: 1.0.0
  contact:
    name: API Support
    url: http://www.dozlab.io/support
    email: support@dozlab.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.dozlab.io
    description: Production server
paths:
...
  /api/v1/lab-sessions:
    post:
      summary: Create lab session
      description: Create a new lab session using Kubernetes CRD
      tags:
        - Lab Sessions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabSessionCreateRequest'
      responses:
        '201':
          description: Lab session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabSessionResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List lab sessions
      description: Get list of user's lab sessions
      tags:
        - Lab Sessions
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Lab sessions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  lab_sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/LabSessionResponse'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
  /api/v1/lab-sessions/{id}:
    get:
      summary: Get lab session
      description: Get specific lab session details
      tags:
        - Lab Sessions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Lab session ID
          schema:
            type: string
      responses:
        '200':
          description: Lab session retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabSessionResponse'
        '404':
          description: Lab session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete lab session
      description: Delete a lab session
      tags:
        - Lab Sessions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Lab session ID
          schema:
            type: string
      responses:
        '200':
          description: Lab session deleted successfully
        '404':
          description: Lab session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LabSessionCreateRequest:
      type: object
      required:
        - lab_id
        - spec_version
      properties:
        lab_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        spec_version:
          type: string
          example: "v1.0.0"
    LabSessionResponse:
      type: object
      properties:
        id:
          type: string
          example: "lab-session-abc123"
        lab_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        user_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        status:
          type: string
          enum: [pending, running, completed, failed]
          example: "running"
        created_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid request format"
        details:
          type: string
          example: "Missing required field: username"
...